agents/requirements.txt
# LiveKit Agent Dependencies
livekit-agents[openai,silero,deepgram,turn-detector]~=1.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
